refactor: atualiza imports para usar novas interfaces
