+++ nexus-backend/src/modules/deliveries/subscribers/notification.subscriber.ts
@@ -2,33 +2,7 @@ import { EntitySubscriberInterface, EventSubscriber, InsertEvent, UpdateEvent } 
 import { Logger, Injectable } from '@nestjs/common';
 import { Delivery } from '../entities/delivery.entity';
 import { DeliveryStatus } from '../enums/delivery-status.enum';
-
-/**
- * Tipos de notificação
- */
-export enum NotificationType {
-  STATUS_CHANGE = 'STATUS_CHANGE',
-  DELIVERY_CREATED = 'DELIVERY_CREATED',
-  DELIVERY_ASSIGNED = 'DELIVERY_ASSIGNED',
-  DELIVERY_COMPLETED = 'DELIVERY_COMPLETED',
-  DELIVERY_FAILED = 'DELIVERY_FAILED',
-  DELIVERY_CANCELLED = 'DELIVERY_CANCELLED',
-}
-
-/**
- * Interface para notificação
- */
-export interface DeliveryNotification {
-  type: NotificationType;
-  deliveryId: string;
-  trackingCode: string;
-  customerId?: string;
-  driverId?: string | undefined;
-  status: DeliveryStatus;
-  previousStatus?: DeliveryStatus;
-  message: string;
-  timestamp: Date;
-}
+import { NotificationType, type DeliveryNotification } from '../interfaces/notification.interface';
 
 /**
  * Subscriber para disparar notificações automáticas
